+++ src/components/Card/Footer/ClearCompletedButton/ClearCompletedButton.tsx
